Move rules and index.js to src
Add a test for index.js
