+++ src/rules/no-vue-dynamic-i18n.js
