+++ src/rules/valid-package-file-require.js
