+++ tests/no-vue-dynamic-i18n.js
@@ -1,4 +1,4 @@
-const rule = require( '../rules/no-vue-dynamic-i18n' );
+const rule = require( '../src/rules/no-vue-dynamic-i18n' );
 const path = require( 'path' );
 const RuleTester = require( 'eslint' ).RuleTester;
